API

OptimalControlProblems

OptimalControlProblems.metadata

OptimalControlProblems.AbstractModelBackend

OptimalControlProblems.JuMPBackend

OptimalControlProblems.OptimalControlBackend

OptimalControlProblems.available_problems

OptimalControlProblems.beam

OptimalControlProblems.beam

OptimalControlProblems.beam

OptimalControlProblems.bioreactor

OptimalControlProblems.bioreactor

OptimalControlProblems.bioreactor

OptimalControlProblems.cart_pendulum

OptimalControlProblems.cart_pendulum

OptimalControlProblems.cart_pendulum

OptimalControlProblems.chain

OptimalControlProblems.chain

OptimalControlProblems.chain

OptimalControlProblems.dielectrophoretic_particle

OptimalControlProblems.dielectrophoretic_particle

OptimalControlProblems.dielectrophoretic_particle

OptimalControlProblems.double_oscillator

OptimalControlProblems.double_oscillator

OptimalControlProblems.double_oscillator

OptimalControlProblems.ducted_fan

OptimalControlProblems.ducted_fan

OptimalControlProblems.ducted_fan

OptimalControlProblems.electric_vehicle

OptimalControlProblems.electric_vehicle

OptimalControlProblems.electric_vehicle

OptimalControlProblems.glider

OptimalControlProblems.glider

OptimalControlProblems.glider

OptimalControlProblems.insurance

OptimalControlProblems.insurance

OptimalControlProblems.insurance

OptimalControlProblems.jackson

OptimalControlProblems.jackson

OptimalControlProblems.jackson

OptimalControlProblems.moonlander

OptimalControlProblems.moonlander

OptimalControlProblems.moonlander

OptimalControlProblems.quadrotor

OptimalControlProblems.quadrotor

OptimalControlProblems.quadrotor

OptimalControlProblems.robbins

OptimalControlProblems.robbins

OptimalControlProblems.robbins

OptimalControlProblems.robot

OptimalControlProblems.robot

OptimalControlProblems.robot

OptimalControlProblems.rocket

OptimalControlProblems.rocket

OptimalControlProblems.rocket

OptimalControlProblems.space_shuttle

OptimalControlProblems.space_shuttle

OptimalControlProblems.space_shuttle

OptimalControlProblems.steering

OptimalControlProblems.steering

OptimalControlProblems.steering

OptimalControlProblems.truck_trailer

OptimalControlProblems.truck_trailer

OptimalControlProblems.truck_trailer

OptimalControlProblems.vanderpol

OptimalControlProblems.vanderpol

OptimalControlProblems.vanderpol

CTModels.control

CTModels.control

CTModels.costate

CTModels.costate

CTModels.state

CTModels.state

CTModels.time_grid

CTModels.time_grid

ExaModels.variable

ExaModels.variable

OptimalControlProblems.metadata — Constant

OptimalControlProblems.metadata::Dict()

Dictionary containing metadata for all available optimal control problems.

The following keys are valid:

name::String: the problem name.

N::Int: the default number of steps.

minimise::Bool: indicates whether the objective function is minimised (true) or maximised (false).

state_name::Vector{String}: names of the state components.

costate_name::Vector{String}: names of the differential constraints to obtain the costate (dual variables associated with the differential constraints).

control_name::Vector{String}: names of the control components.

variable_name::Union{Vector{String},Nothing}: names of the optimisation variables, or nothing if no such variable exists.

final_time::Tuple{Symbol, Union{Float64, Int}}: of the form (type, value_or_index), where:

type is either :fixed or :free.

value_or_index is the index in variable if the final time is free, or its value if it is fixed.

Example

julia> OptimalControlProblems.metadata[:my_problem][:name]
"My Problem"

OptimalControlProblems.AbstractModelBackend — Type

abstract type AbstractModelBackend

Abstract type for all optimal control problem model back-ends.

OptimalControlProblems.JuMPBackend — Type

struct JuMPBackend <: OptimalControlProblems.AbstractModelBackend

Back-end for modelling optimal control problems using the JuMP optimisation framework.

OptimalControlProblems.OptimalControlBackend — Type

struct OptimalControlBackend <: OptimalControlProblems.AbstractModelBackend

Back-end for modelling optimal control problems using the OptimalControl.jl package.

CTModels.control — Method

control(
    _::Symbol,
    model
) -> Union{JuMPModels.var"#671#673", JuMPModels.var"#672#674"}

Retrieve the control trajectory from a JuMP model.

Arguments

::Symbol: Problem name.

model: JuMP model object.

Returns

Throws ExtensionError(:JuMP) since JuMP support must be extended.

Example

julia> control(:problem1, model)
ERROR: ExtensionError(:JuMP)

CTModels.costate — Method

costate(
    _::Symbol,
    model
) -> Union{JuMPModels.var"#676#679", JuMPModels.var"#677#680"}

Retrieve the costate (adjoint variables) from a JuMP model.

Arguments

::Symbol: Problem name.

model: JuMP model object.

Returns

Throws ExtensionError(:JuMP) since JuMP support must be extended.

Example

julia> costate(:problem1, model)
ERROR: ExtensionError(:JuMP)

CTModels.state — Method

state(
    _::Symbol,
    model
) -> Union{JuMPModels.var"#667#669", JuMPModels.var"#668#670"}

Retrieve the state trajectory from a JuMP model.

Arguments

::Symbol: Problem name.

model: JuMP model object.

Returns

Throws ExtensionError(:JuMP) since JuMP support must be extended.

Example

julia> state(:problem1, model)
ERROR: ExtensionError(:JuMP)

CTModels.time_grid — Method

time_grid(_::Symbol, model) -> Any

Retrieve the discretised time grid from a JuMP model.

Arguments

::Symbol: Problem name.

model: JuMP model object.

Returns

Throws ExtensionError(:JuMP) since JuMP support must be extended.

Example

julia> time_grid(:problem1, model)
ERROR: ExtensionError(:JuMP)

ExaModels.variable — Method

variable(_::Symbol, model) -> Any

Retrieve optimisation variables from a JuMP model.

Arguments

::Symbol: Problem name.

model: JuMP model object.

Returns

Throws ExtensionError(:JuMP) since JuMP support must be extended.

Example

julia> variable(:problem1, model)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.available_problems — Method

available_problems() -> Vector{Symbol}

Returns the list of available optimal control problems.

Returns

Vector{Symbol}: A vector of problem names as symbols.

Example

julia> OptimalControlProblems.available_problems()
[:problem1, :problem2, :problem3]

OptimalControlProblems.beam — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem beam for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> beam(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.bioreactor — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem bioreactor for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> bioreactor(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.cart_pendulum — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem cart_pendulum for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> cart_pendulum(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.chain — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem chain for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> chain(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.dielectrophoretic_particle — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem dielectrophoretic_particle for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> dielectrophoretic_particle(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.double_oscillator — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem double_oscillator for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> double_oscillator(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.ducted_fan — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem ducted_fan for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> ducted_fan(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.electric_vehicle — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem electric_vehicle for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> electric_vehicle(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.glider — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem glider for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> glider(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.insurance — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem insurance for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> insurance(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.jackson — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem jackson for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> jackson(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.moonlander — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem moonlander for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> moonlander(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.quadrotor — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem quadrotor for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> quadrotor(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.robbins — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem robbins for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> robbins(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.robot — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem robot for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> robot(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.rocket — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem rocket for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> rocket(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.space_shuttle — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem space_shuttle for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> space_shuttle(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.steering — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem steering for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> steering(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.truck_trailer — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem truck_trailer for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> truck_trailer(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlProblems.vanderpol — Method

DocStringExtensions.TypedMethodSignatures()

Defines the optimal control problem vanderpol for a given back-end.

Arguments

model_backend::T: The modelling back-end, subtype of AbstractModelBackend.

N::Int=0: Number of discretisation steps (optional).

Returns

Throws an ExtensionError if the required back-end is not available.

Example

julia> vanderpol(JuMPBackend(); N=20)
ERROR: ExtensionError(:JuMP)

OptimalControlModels

OptimalControlProblems.beam — Method

beam(
    ::OptimalControlBackend;
    N
) -> Tuple{CTDirect.DOCP{_A, CTModels.Model{TD, TimesModelType, StateModelType, ControlModelType, VariableModelType, DynamicsModelType, ObjectiveModelType, ConstraintsModelType, BuildExaModelType}} where {_A<:CTDirect.Discretization, TD<:CTModels.TimeDependence, TimesModelType<:CTModels.AbstractTimesModel, StateModelType<:CTModels.AbstractStateModel, ControlModelType<:CTModels.AbstractControlModel, VariableModelType<:CTModels.AbstractVariableModel, DynamicsModelType<:Function, ObjectiveModelType<:CTModels.AbstractObjectiveModel, ConstraintsModelType<:CTModels.AbstractConstraintsModel, BuildExaModelType<:Union{Nothing, Function}}, Any}

Constructs an OptimalControl problem representing the Beam problem using the OptimalControl backend, as formulated in the BOCOP repository. The model includes state and control variables, boundary conditions, dynamics, path constraints, and an objective functional.

Arguments

::OptimalControlBackend: Placeholder for the OptimalControl backend type, selecting the solver interface.

N::Int=500: Number of discretisation points for the direct transcription.

Returns

docp: The direct optimal control problem (DOCP) object.

nlp: The corresponding nonlinear programming (NLP) model obtained from direct transcription.

Example

julia> docp, nlp = OptimalControlProblems.beam(OptimalControlBackend(); N=500)
(DOCP object, NLP model)

OptimalControlProblems.bioreactor — Method

The Bioreactor Problem: The problem is formulated as an OptimalControl model and can be found here

OptimalControlProblems.cart_pendulum — Method

The Cart-Pendulum Problem:
    we want to find the optimal trajectory of a cart-pendulum system.
    The objective is to swing the pendulum from the downward position to the upright position in the shortest time possible.
    The problem is formulated as an OptimalControl model.

OptimalControlProblems.chain — Method

The Hanging Chain Problem: We want to find the shape of a chain hanging between two points a and b, with a length L. The objective is to minimise the potential energy of the chain. The problem is formulated as an OptimalControl model.

OptimalControlProblems.dielectrophoretic_particle — Method

Dielectrophoretic particle problem: This problem consists of a dielectrophoretic particle system. The goal is to find the trajectory that minimise the time taken for the particle to travel between two points. The problem is formulated as an OptimalControl model. Ref: [CPR2006] Chang, D. E., Petit, N., & Rouchon, P. (2006). Time-optimal control of a particle in a dielectrophoretic system. IEEE Transactions on Automatic Control, 51(7), 1100-1114.

OptimalControlProblems.double_oscillator — Method

Double Oscillator Problem: Implement the optimal control of a double oscillator toy model. The problem is formulated as an OptimalControl model. Ref: [CLP2018] Coudurier, C., Lepreux, O., & Petit, N. (2018). Optimal bang-bang control of a mechanical double oscillator using averaging methods. IFAC-PapersOnLine, 51(2), 49-54.

OptimalControlProblems.ducted_fan — Method

The Ducted Fan Problem: Implement the optimal control of a planar ducted fan. Instance taken from [GP2009]. The problem is formulated as an OptimalControl model. Ref: Graichen, K., & Petit, N. (2009). Incorporating a class of constraints into the dynamics of optimal control problems. Optimal Control Applications and Methods, 30(6), 537-561.

OptimalControlProblems.electric_vehicle — Method

The electric Vehicle Problem Implement optimal control of an electric vehicle. The problem is formulated as an OptimalControl model. Ref: [PS2011] Nicolas Petit and Antonio Sciarretta. "Optimal drive of electric vehicles using an inversion-based trajectory generation approach." IFAC Proceedings Volumes 44, no. 1 (2011): 14519-14526.

OptimalControlProblems.glider — Method

Hang Glider Problem: We want to find the optimal trajectory of a hang glider. The objective is to maximize the final horizontal position of the glider while in the presence of a thermal updraft. The problem is formulated as an OptimalControl model. Original formulation from MadNLP/COPSBenchmark

OptimalControlProblems.insurance — Method

The Insurance Problem: The problem is formulated as an OptimalControl model and can be found here

OptimalControlProblems.jackson — Method

The Jackson Problem: The problem is formulated as an OptimalControl model and can be found here

OptimalControlProblems.moonlander — Method

The Moonlander Problem: We want to find the optimal trajectory for a moonlander to land on the moon. The objective is to minimise the time taken to land on the moon. The problem is formulated as an OptimalControl model.

OptimalControlProblems.quadrotor — Method

Quadrotor Problem: We want to find the optimal trajectory of a quadrotor to reach a target position. The objective is to minimise the final time. The problem is formulated as an OptimalControl model.

OptimalControlProblems.robbins — Method

The Robbins Problem: The problem is formulated as an OptimalControl model and can be found here

OptimalControlProblems.robot — Method

Robot arm problem: We want to find the shape of a robot arm moving between two points. The objective is to minimise the time taken to move between the two points. The problem is formulated as an OptimalControl model.

OptimalControlProblems.rocket — Method

Goddard Rocket Problem: We want to find the optimal trajectory of a Goddard rocket. The objective is to maximize the final altitude of the rocket. The problem is formulated as an OptimalControl model.

OptimalControlProblems.space_shuttle — Method

Space Shuttle Reentry Trajectory Problem: We want to find the optimal trajectory of a space shuttle reentry. The objective is to maximize the latitude (cross range) at the terminal point. The original problem formulated as a JuMP model can be found here Note: no heating limit path constraint

OptimalControlProblems.steering — Method

Particle Steering Problem: We want to find the optimal trajectory of a particle. The objective is to minimise the time taken to achieve a given altitude and terminal velocity. The problem is formulated as an OptimalControl model.

OptimalControlProblems.truck_trailer — Method

The Truck Trailer Problem: We want to find the optimal trajectory of a truck with two trailers that starts horizontally aligned. The objective is to minimise the time taken to park the truck and the trailers aligned vertically at a given target location. The problem is formulated as an OptimalControl model.

OptimalControlProblems.vanderpol — Method

The Van der Pol Problem: The problem is formulated as an OptimalControl model and can be found here

JuMPModels

CTModels.control

CTModels.control

CTModels.costate

CTModels.costate

CTModels.state

CTModels.state

CTModels.time_grid

CTModels.time_grid

ExaModels.variable

ExaModels.variable

CTModels.control — Method

control(
    problem::Symbol,
    model::GenericModel
) -> Union{JuMPModels.var"#671#673", JuMPModels.var"#672#674"}

Extract and interpolate the control trajectory from a JuMP model of an optimal control problem.

Arguments

problem::Symbol: The name of the problem as defined in OptimalControlProblems.metadata.

model::JuMP.GenericModel: The JuMP model containing the problem solution.

Returns

fu::Function: A function of continuous time returning the interpolated control. If the control is scalar, the function returns a scalar; otherwise, a vector.

Example

julia> u = OptimalControlProblems.control(:my_problem, model)
julia> u(0.25)
0.42

CTModels.costate — Method

costate(
    problem::Symbol,
    model::GenericModel
) -> Union{JuMPModels.var"#676#679", JuMPModels.var"#677#680"}

Extract and interpolate the costate trajectory (dual variables associated with states) from a JuMP model of an optimal control problem.

Arguments

problem::Symbol: The name of the problem as defined in OptimalControlProblems.metadata.

model::JuMP.GenericModel: The JuMP model containing the problem solution.

Returns

fp::Function: A function of continuous time returning the interpolated costate. If the costate is scalar, the function returns a scalar; otherwise, a vector.

Example

julia> p = OptimalControlProblems.costate(:my_problem, model)
julia> p(0.75)
[-0.12, 0.05]

CTModels.state — Method

state(
    problem::Symbol,
    model::GenericModel
) -> Union{JuMPModels.var"#667#669", JuMPModels.var"#668#670"}

Extract and interpolate the state trajectory from a JuMP model of an optimal control problem.

Arguments

problem::Symbol: The name of the problem as defined in OptimalControlProblems.metadata.

model::JuMP.GenericModel: The JuMP model containing the problem solution.

Returns

fx::Function: A function of continuous time returning the interpolated state. If the state is scalar, the function returns a scalar; otherwise, a vector.

Example

julia> x = OptimalControlProblems.state(:my_problem, model)
julia> x(0.5)
[0.23, 0.71]

CTModels.time_grid — Method

time_grid(problem::Symbol, model::GenericModel) -> Any

Compute the discretised time grid for a given optimal control problem solved with JuMP.

Arguments

problem::Symbol: The name of the problem as defined in OptimalControlProblems.metadata.

model::JuMP.GenericModel: The JuMP model containing the problem solution.

Returns

t_jp::AbstractVector{Float64}: A vector of time points spanning from initial time t0 = 0 to the final time tf.

Example

julia> tgrid = OptimalControlProblems.time_grid(:my_problem, model)
0.0:0.1:1.0

ExaModels.variable — Method

variable(problem::Symbol, model::GenericModel) -> Any

Extract scalar or vector decision variables (such as final time when free) from a JuMP model of an optimal control problem.

Arguments

problem::Symbol: The name of the problem as defined in OptimalControlProblems.metadata.

model::JuMP.GenericModel: The JuMP model containing the problem solution.

Returns

var::Union{Nothing,Float64,Vector{Float64}}:

nothing if the problem defines no additional variables.

A scalar if there is one variable.

A vector if multiple variables exist.

Example

julia> v = OptimalControlProblems.variable(:my_problem, model)
1.5

OptimalControlProblems.beam — Method

beam(::JuMPBackend; N) -> JuMP.Model

Constructs a JuMP model representing the Beam optimal control problem, as formulated in the BOCOP repository.

Arguments

::JuMPBackend: Placeholder for JuMP backend type; used to select the solver interface.

N::Int=500: Number of discretisation steps for the time grid.

Returns

model::JuMP.Model: A JuMP model containing the variables, constraints, dynamics, and objective of the beam problem.

Example

julia> model = OptimalControlProblems.beam(JuMPBackend(); N=500)
# JuMP model with 3*(N+1) variables and corresponding constraints

OptimalControlProblems.bioreactor — Method

The Bioreactor Problem: The problem is formulated as a JuMP model and can be found here

OptimalControlProblems.cart_pendulum — Method

The Cart-Pendulum Problem: we want to find the optimal trajectory of a cart-pendulum system. The objective is to swing the pendulum from the downward position to the upright position in the shortest time possible. The problem is formulated as a JuMP model, and can be found here.

OptimalControlProblems.chain — Method

The Hanging Chain Problem: We want to find the shape of a chain hanging between two points a and b, with a length L. The objective is to minimise the potential energy of the chain. The problem is formulated as a JuMP model, and can be found here

OptimalControlProblems.dielectrophoretic_particle — Method

Dielectrophoretic particle problem: This problem consists of a dielectrophoretic particle system. The goal is to find the trajectory that minimise the time taken for the particle to travel between two points. The problem is formulated as a JuMP model. Ref: [CPR2006] Chang, D. E., Petit, N., & Rouchon, P. (2006). Time-optimal control of a particle in a dielectrophoretic system. IEEE Transactions on Automatic Control, 51(7), 1100-1114.

OptimalControlProblems.double_oscillator — Method

Double Oscillator Problem: Implement the optimal control of a double oscillator toy model. The problem is formulated as a JuMP model. Ref: [CLP2018] Coudurier, C., Lepreux, O., & Petit, N. (2018). Optimal bang-bang control of a mechanical double oscillator using averaging methods. IFAC-PapersOnLine, 51(2), 49-54.

OptimalControlProblems.ducted_fan — Method

The Ducted Fan Problem: Implement the optimal control of a planar ducted fan. Instance taken from [GP2009]. The problem is formulated as a JuMP model. Ref: Graichen, K., & Petit, N. (2009). Incorporating a class of constraints into the dynamics of optimal control problems. Optimal Control Applications and Methods, 30(6), 537-561.

OptimalControlProblems.electric_vehicle — Method

The electric Vehicle Problem Implement optimal control of an electric vehicle. The problem is formulated as a JuMP model. Ref: [PS2011] Nicolas Petit and Antonio Sciarretta. "Optimal drive of electric vehicles using an inversion-based trajectory generation approach." IFAC Proceedings Volumes 44, no. 1 (2011): 14519-14526.

OptimalControlProblems.glider — Method

Hang Glider Problem: We want to find the optimal trajectory of a hang glider. The objective is to maximize the final horizontal position of the glider while in the presence of a thermal updraft. The problem is formulated as a JuMP model, and can be found here

OptimalControlProblems.insurance — Method

The Insurance Problem: The problem is formulated as a JuMP model and can be found here

OptimalControlProblems.jackson — Method

The Jackson Problem: The problem is formulated as a JuMP model and can be found here

OptimalControlProblems.moonlander — Method

The Moonlander Problem: We want to find the optimal trajectory for a moonlander to land on the moon. The objective is to minimise the time taken to land on the moon. The problem is formulated as a JuMP model, and can be found here

OptimalControlProblems.quadrotor — Method

Quadrotor Problem: We want to find the optimal trajectory of a quadrotor to reach a target position. The objective is to minimise the final time. The problem is formulated as a JuMP model, and can be found here

OptimalControlProblems.robbins — Method

The Robbins Problem: The problem is formulated as a JuMP model and can be found here

OptimalControlProblems.robot — Method

Robot arm problem: We want to find the shape of a robot arm moving between two points. The objective is to minimise the time taken to move between the two points. The problem is formulated as a JuMP model, and can be found here

OptimalControlProblems.rocket — Method

Goddard Rocket Problem: We want to find the optimal trajectory of a Goddard rocket. The objective is to maximize the final altitude of the rocket. The problem is formulated as a JuMP model, and can be found here

OptimalControlProblems.space_shuttle — Method

Space Shuttle Reentry Trajectory Problem: We want to find the optimal trajectory of a space shuttle reentry. The objective is to maximize the latitude (cross range) at the terminal point. The problem is formulated as a JuMP model, and can be found here Note: no heating limit path constraint

OptimalControlProblems.steering — Method

Particle Steering Problem: We want to find the optimal trajectory of a particle. The objective is to minimise the time taken to achieve a given altitude and terminal velocity. The problem is formulated as a JuMP model, and can be found here

OptimalControlProblems.truck_trailer — Method

The Truck Trailer Problem: We want to find the optimal trajectory of a truck with two trailers that starts horizontally aligned. The objective is to minimise the time taken to park the truck and the trailers aligned vertically at a given target location. The problem is formulated as a JuMP model, and can be found here

OptimalControlProblems.vanderpol — Method

The Van der Pol Problem: The problem is formulated as a JuMP model and can be found here